change (1.0, springbone): move colliders to spring bone module (1)
