--- packages/three-vrm-node-collider/README.md
@@ -1,5 +1,0 @@
-# @pixiv/three-vrm-node-collider
-
-Node collider module for @pixiv/three-vrm
-
-You cannot use this yet! It's a candidate implementation for VRM 1.0
