--- packages/three-vrm-node-collider/jest.config.js
@@ -1,19 +1,0 @@
-module.exports = {
-  "roots": [
-    "<rootDir>/src"
-  ],
-  "transform": {
-    "^.+\\.tsx?$": "ts-jest"
-  },
-  "testRegex": "[^.]*.test.ts$",
-  "moduleFileExtensions": [
-    "ts",
-    "js",
-    "json",
-    "node"
-  ],
-  "collectCoverageFrom": [
-    "**/*.ts",
-    "!**/tests/**"
-  ]
-}
