--- packages/three-vrm-node-collider/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@pixiv/three-vrm-node-collider",
-  "version": "0.0.1",
-  "description": "Node collider module for @pixiv/three-vrm",
-  "license": "MIT",
-  "author": "pixiv",
-  "files": [
-    "/lib/",
-    "/ts*/",
-    "/types/",
-    "LICENSE"
-  ],
-  "main": "lib/three-vrm-node-collider.min.js",
-  "types": "types/index.d.ts",
-  "typesVersions": {
-    "<3.9": {
-      "*": [
-        "ts3.4/*"
-      ]
-    }
-  },
-  "scripts": {
-    "version": "yarn all",
-    "all": "yarn && yarn lint && yarn clean && yarn build && yarn docs",
-    "dev": "cross-env NODE_ENV=development SERVE=1 rollup -w -c -o lib/three-vrm-node-collider.js",
-    "clean": "rimraf docs/ lib/ ts*/ types/",
-    "build": "yarn build-dev && yarn build-prod && yarn build-moduledev && yarn build-moduleprod && yarn build-types",
-    "build-dev": "cross-env NODE_ENV=development rollup -c -o lib/three-vrm-node-collider.js",
-    "build-prod": "cross-env NODE_ENV=production rollup -c -o lib/three-vrm-node-collider.min.js",
-    "build-moduledev": "cross-env NODE_ENV=development ESM=1 rollup -c -o lib/three-vrm-node-collider.module.js",
-    "build-moduleprod": "cross-env NODE_ENV=production ESM=1 rollup -c -o lib/three-vrm-node-collider.module.min.js",
-    "build-types": "tsc --declaration --declarationDir ./types --emitDeclarationOnly && downlevel-dts types ts3.4/types",
-    "docs": "typedoc --out docs --mode file --excludeNotExported",
-    "test": "jest",
-    "lint": "eslint \"src/**/*.{ts,tsx}\" && prettier \"src/**/*.{ts,tsx}\" --check",
-    "lint-fix": "eslint \"src/**/*.{ts,tsx}\" --fix && prettier \"src/**/*.{ts,tsx}\" --write"
-  },
-  "lint-staged": {
-    "./src/**/*.{ts,tsx}": [
-      "eslint --fix",
-      "prettier --write"
-    ]
-  },
-  "dependencies": {
-    "@pixiv/types-vrm-0.0": "0.0.1"
-  },
-  "devDependencies": {
-    "lint-staged": "10.4.2",
-    "three": "^0.125.2"
-  },
-  "peerDependencies": {
-    "three": "^0.125.2"
-  }
-}
