--- packages/three-vrm-node-collider/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export { VRMNodeColliderHelper } from './VRMNodeColliderHelper';
