--- packages/three-vrm-node-collider/src/index.ts
@@ -1,6 +1,0 @@
-export * from './helpers';
-export * from './VRMNodeCollider';
-export * from './VRMNodeColliderImporter';
-export * from './VRMNodeColliderShape';
-export * from './VRMNodeColliderShapeCapsule';
-export * from './VRMNodeColliderShapeSphere';
