+++ packages/three-vrm-springbone/src/VRMSpringBoneColliderShape.ts
