+++ packages/three-vrm-springbone/src/helpers/utils/ColliderShapeBufferGeometry.ts
